Convert finalist scheduling to use JS-joda time and the time picker
This makes it easier for the user to choose start and end times on the
parameters page. The use of JS-joda makes the javascript much more
compatible with the Java time objects.

In doing this I ended up replacing jStorage with a custom wrapper that
properly serializes and deserializes the JS-joda time objects that are
used.

Issue #447
Fixes #909