Make loading and saving to local storage explicit
The finalist object no longer automatically loads from local storage or
saves to local storage. This allows the calling page to determine when
(or if) the information should be saved to local storage.

Also fixed an error in the schedule.js where moving teams between cells
caused an invalid schedule object to be created.

Issue #447