WIP: prompt user for scheduler columns in the GUI
The GUI and the web are to the same point now.

- Moved construction of ChooseChallengeDescriptor from the SchedulerUI
constructor to the method where it's used as this causes the dialog
to be recognized as a utility window rather than it's own window.

- SchedulerUI now has a current challenge description that is
persistent. This can be leveraged for later changes to the
automatic scheduler and not need to prompt the user for some
information.